+++ src/main/java/com/foolsix/fancyenchantments/enchantment/EaterOfSouls.java
@@ -1,12 +1,9 @@
 package com.foolsix.fancyenchantments.enchantment;
 
 import com.foolsix.fancyenchantments.FancyEnchantments;
-import com.foolsix.fancyenchantments.enchantment.EssentiaEnch.FEBaseEnchantment;
+import com.foolsix.fancyenchantments.enchantment.EssentiaEnch.TwistedEnchantment;
 import com.foolsix.fancyenchantments.util.ModConfig;
-import net.minecraft.ChatFormatting;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.MobType;
 import net.minecraft.world.entity.player.Player;
@@ -14,30 +11,24 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
 
-public class EaterOfSouls extends FEBaseEnchantment {
+public class EaterOfSouls extends TwistedEnchantment {
     private static final ModConfig.EaterOfSoulsOptions CONFIG = FancyEnchantments.getConfig().eaterOfSoulsOptions;
+    private static final String KILL_COUNT = "eater_of_souls_killcount";
 
     public EaterOfSouls() {
         super(CONFIG, EnchantmentCategory.WEAPON, new EquipmentSlot[]{EquipmentSlot.MAINHAND});
     }
 
-    @Override
-    public Component getFullname(int level) {
-        return ((MutableComponent) super.getFullname(level)).withStyle(ChatFormatting.DARK_PURPLE);
-    }
-
     public void killcount(LivingDeathEvent e) {
         if (e.getSource().getEntity() instanceof Player player) {
             ItemStack tool = player.getMainHandItem();
+            //On main hand to get kill count
             if (tool.getEnchantmentLevel(this) > 0) {
                 e.getEntity().skipDropExperience();
                 CompoundTag nbt = tool.getOrCreateTag();
-                if (nbt.contains("eater_of_souls_killcount")) {
-                    nbt.putInt("eater_of_souls_killcount", nbt.getInt("eater_of_souls_killcount") + 1);
-                } else {
-                    nbt.putInt("eater_of_souls_killcount", 1);
-                }
+                nbt.putInt(KILL_COUNT, nbt.getInt(KILL_COUNT) + 1);
             }
+            //On offhand still cancel drop experience
             if (player.getOffhandItem().getEnchantmentLevel(this) > 0) {
                 e.getEntity().skipDropExperience();
             }
@@ -47,10 +38,7 @@ public class EaterOfSouls extends FEBaseEnchantment {
     @Override
     public float getDamageBonus(int level, MobType mobType, ItemStack enchantedItem) {
         if (level > 0) {
-            CompoundTag nbt = enchantedItem.getTag();
-            if (nbt != null && nbt.contains("eater_of_souls_killcount")) {
-                return (float) (CONFIG.damageMultiplier * Math.sqrt(nbt.getInt("eater_of_souls_killcount")));
-            }
+            return CONFIG.damageMultiplier * enchantedItem.getOrCreateTag().getInt(KILL_COUNT);
         }
         return 0;
     }